+++ web/ui/react-app/src/config/globals.ts
