+++ web/ui/react-app/src/pages/flags/Flags.tsx
@@ -1,9 +1,9 @@
 import React, { FC } from 'react';
 import { RouteComponentProps } from '@reach/router';
 import { Table } from 'reactstrap';
-import { withStatusIndicator } from '../withStatusIndicator';
-import { useFetch } from '../utils/useFetch';
-import PathPrefixProps from '../PathPrefixProps';
+import { withStatusIndicator } from '../../common/withStatusIndicator';
+import { useFetch } from '../../common/hooks/useFetch';
+import PathPrefixProps from '../../PathPrefixProps';
 
 interface FlagMap {
   [key: string]: string;
