+++ web/ui/react-app/src/pages/graph/QueryStatsView.css
