+++ web/ui/react-app/src/pages/graph/SeriesName.test.tsx
