+++ web/ui/react-app/src/pages/index.ts
@@ -1,11 +1,11 @@
 import Alerts from './alerts/Alerts';
-import Config from './Config';
-import Flags from './Flags';
-import Rules from './Rules';
-import Services from './Services';
-import Status from './Status';
+import Config from './config/Config';
+import Flags from './flags/Flags';
+import Rules from './rules/Rules';
+import Services from './serviceDiscovery/Services';
+import Status from './status/Status';
 import Targets from './targets/Targets';
-import PanelList from './PanelList';
-import TSDBStatus from './TSDBStatus';
+import PanelList from './graph/PanelList';
+import TSDBStatus from './tsdbStatus/TSDBStatus';
 
 export { Alerts, Config, Flags, Rules, Services, Status, Targets, TSDBStatus, PanelList };
