+++ web/ui/react-app/src/pages/status/Status.tsx
@@ -1,9 +1,9 @@
 import React, { Fragment, FC } from 'react';
 import { RouteComponentProps } from '@reach/router';
 import { Table } from 'reactstrap';
-import { withStatusIndicator } from '../withStatusIndicator';
-import { useFetch } from '../utils/useFetch';
-import PathPrefixProps from '../PathPrefixProps';
+import { withStatusIndicator } from '../../common/withStatusIndicator';
+import { useFetch } from '../../common/hooks/useFetch';
+import PathPrefixProps from '../../PathPrefixProps';
 
 const sectionTitles = ['Runtime Information', 'Build Information', 'Alertmanagers'];
 
@@ -65,7 +65,7 @@ export const StatusContent: FC<StatusPageProps> = ({ data = [] }) => {
             <h2>{sectionTitles[i]}</h2>
             <Table className="h-auto" size="sm" bordered striped>
               <tbody>
-                {Object.entries(statuses).map(([k, v], i) => {
+                {Object.entries(statuses).map(([k, v]) => {
                   const { title = k, customizeValue = (val: any) => val, customRow, skip } = statusConfig[k] || {};
                   if (skip) {
                     return null;
