+++ web/ui/react-app/src/utils/MetricFormat.test.ts
